+++ Shared/ApplianceViewHelpers.swift
@@ -19,3 +19,14 @@ func getDeviceIcon(battery: Battery) -> Image {
         return Image(systemName: "iphone")
     }
 }
+
+func carDetails(car: Car) -> String {
+    var text = ""
+    if car.vehicle.engine { text += "Car on | " }
+
+    if car.vehicle.hvac { text += "Climate on | " }
+
+    if car.vehicle.distance != 0 { text += "Range \(car.vehicle.distance) km | " }
+    if car.vehicle.evStatusTimestamp != "" { text += "Updated \(getCarTime(strDate: car.vehicle.evStatusTimestamp))" }
+    return text
+}
